Capabilities

+++ crates/font-mud/examples/pack_mononoki.rs
@@ -22,7 +22,7 @@ fn main() {
     let (glyph_atlas, _glyph_shape_errors) = GlyphAtlas::new(&face).unwrap();
     let bitmap = &glyph_atlas.bitmap;
     let output = File::create("mononoki.png").unwrap();
-    let mut encoder = png::Encoder::new(output, bitmap.width as u32, bitmap.height as u32);
+    let mut encoder = png::Encoder::new(output, bitmap.width, bitmap.height);
     encoder.set_color(png::ColorType::Rgba);
     encoder.set_depth(png::BitDepth::Eight);
     let mut writer = encoder.write_header().unwrap();
